pyproject.toml
[project]
name = "cineverse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.121.2",
    "flask>=3.1.2",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.4",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.44",
    "typing>=3.10.0.0",
    "uuid>=1.30",
]
